feat: validation and appropriate response
json data from the request body is checked for login key and password, if both present, login and password are validated.
if the validation passes, the data is added to the db and a response is sent with status code 204 and the message "Registering successfull". If not, a response with status code 404 and the message "Login or password invalid" appears.