feat:Authentication, unique login in db
When attempting to register, it is checked whether the user has entered the correct data and whether a user with the same login already exists. When trying to log in, it is checked whether the user exists, and then the data entered by the user is compared with the data from the database. If the data is true a JWT is returned.
 All possibilities are checked, such as not entering a password, entering bad data, etc., and corresponding responses with status codes are returned