feat: authorized endpoint /items added, new class item created
The http request must have an authorized header with the value "Bearer + token". In the body of the request is sent the name of the item to be added to the user who owns the token. In case the key is wrong, expired or there is no authorized header, the server responds with a status code of 401 and a message "Unauthorized". Otherwise, server decodes token to get name of the user and adds item for this particular user.